Move HybridEP tests to normal suite
