+++ tests/single_card_tests/ai_edited_test/moe/test_ai_token_dispatcher.py
@@ -99,119 +99,6 @@ class TestTokenDispatcher(unittest.TestCase):
             self.assertIsNotNone(dispatcher._comm_manager)
             self.assertEqual(dispatcher.num_local_experts, 2)
 
-    def test_deep_ep_backend_defaults_to_deepep_and_ignores_env(self):
-        """Test backend selection is controlled by config, not env vars."""
-        from paddlefleet.transformer.moe.token_dispatcher import (
-            get_selected_deep_ep_backend_name,
-        )
-
-        with (
-            patch.dict(
-                os.environ,
-                {
-                    "PF_DEEP_EP_BACKEND": "hybrid",
-                    "PADDLEFLEET_DEEP_EP_BACKEND": "hybrid",
-                },
-            ),
-            patch(
-                "paddlefleet.transformer.moe.token_dispatcher.HAVE_HYBRID_EP",
-                True,
-            ),
-        ):
-            self.assertEqual(get_selected_deep_ep_backend_name(), "deepep")
-
-    def test_deep_ep_backend_selects_hybrid_explicitly(self):
-        """Test explicit config selects HybridEP backend."""
-        from paddlefleet.transformer.moe.token_dispatcher import (
-            get_selected_deep_ep_backend_name,
-            is_hybrid_ep_backend_selected,
-        )
-
-        with patch(
-            "paddlefleet.transformer.moe.token_dispatcher.HAVE_HYBRID_EP",
-            True,
-        ):
-            self.assertEqual(
-                get_selected_deep_ep_backend_name("hybridep"), "hybrid"
-            )
-            self.assertTrue(is_hybrid_ep_backend_selected("hybrid_ep"))
-
-    def test_deep_ep_backend_deep_ep_v2_is_reserved(self):
-        """Test reserved deep_ep_v2 backend is explicit but unsupported."""
-        from paddlefleet.transformer.moe.token_dispatcher import (
-            get_selected_deep_ep_backend_name,
-        )
-
-        with self.assertRaisesRegex(NotImplementedError, "deep_ep_v2"):
-            get_selected_deep_ep_backend_name("deep_ep_v2")
-
-    def test_moe_flex_token_dispatcher_uses_hybrid_backend(self):
-        """Test MoEFlexTokenDispatcher chooses HybridEP from config."""
-        from paddlefleet.transformer.moe.token_dispatcher import (
-            MoEFlexTokenDispatcher,
-        )
-
-        mock_group = MagicMock()
-        mock_group.world_size = 2
-        mock_manager_cls = MagicMock()
-
-        with (
-            patch(
-                "paddlefleet.transformer.moe.token_dispatcher.HAVE_HYBRID_EP",
-                True,
-            ),
-            patch(
-                "paddlefleet.transformer.moe.token_dispatcher._HybridEPManager",
-                mock_manager_cls,
-            ),
-        ):
-            dispatcher = MoEFlexTokenDispatcher(
-                num_local_experts=2,
-                num_experts_per_tok=2,
-                n_routed_experts=4,
-                ep_group=mock_group,
-                backend_name="hybridep",
-                needs_host_counts=True,
-            )
-            self.assertIs(
-                dispatcher._comm_manager, mock_manager_cls.return_value
-            )
-            self.assertEqual(
-                mock_manager_cls.call_args.kwargs["needs_host_counts"], True
-            )
-
-    def test_hybrid_ep_manager_combine_rejects_overlap(self):
-        """Test HybridEP combine rejects fake overlap handles."""
-        from paddlefleet.transformer.moe.token_dispatcher import (
-            _HybridEPManager,
-        )
-
-        manager = _HybridEPManager.__new__(_HybridEPManager)
-        with self.assertRaisesRegex(NotImplementedError, "combine overlap"):
-            manager.combine(paddle.zeros([1, 4]), {"fn": MagicMock()})
-
-    def test_hybrid_ep_manager_combine_clears_runtime_state(self):
-        """Test HybridEP combine calls the sync bridge and clears state."""
-        from paddlefleet.transformer.moe.token_dispatcher import (
-            _HybridEPManager,
-        )
-
-        manager = _HybridEPManager.__new__(_HybridEPManager)
-        manager.dispatched_probs = object()
-        manager.handle = object()
-        hidden = paddle.zeros([1, 4])
-        combined = object()
-        with patch(
-            "paddlefleet.transformer.moe.token_dispatcher.hybrid_ep_combine",
-            return_value=combined,
-        ) as mock_combine:
-            result = manager.combine(hidden)
-
-        self.assertIs(result, combined)
-        self.assertIsNone(manager.dispatched_probs)
-        self.assertIsNone(manager.handle)
-        mock_combine.assert_called_once_with(hidden, manager)
-
     def test_alltoall_token_dispatcher_init(self):
         """Test AllToAllTokenDispatcher initialization."""
         from paddlefleet.transformer.moe.token_dispatcher import (
